Fixed the terminology flow and made more translations

+++ TKDprogress_BLL/Services/TerminologyService.cs
@@ -27,13 +27,35 @@ namespace TKDprogress_BLL.Services
 
         public async Task<TerminologyDto> GetTerminologyByIdAsync(int id)
         {
-            TerminologyDto artist = await _terminologyRepository.GetTerminologyByIdAsync(id);
+            if (id <= 0)
+            {
+                return new TerminologyDto { ErrorMessage = "Invalid terminology." };
+            }
 
-            return artist;
+            TerminologyDto terminology = await _terminologyRepository.GetTerminologyByIdAsync(id);
+
+            if (terminology.Id <= 0)
+            {
+                return new TerminologyDto { ErrorMessage = "Terminology not found." };
+            }
+
+            return terminology;
         }
 
         public async Task<TerminologyDto> CreateTerminologyAsync(TerminologyDto terminology)
         {
+            if (terminology.CategoryId <= 0)
+            {
+                terminology.ErrorMessage = "Category type does not exist.";
+                return terminology;
+            }
+
+            if (string.IsNullOrEmpty(terminology.Word) || string.IsNullOrEmpty(terminology.Meaning))
+            {
+                terminology.ErrorMessage = "Terminology name or description has an incorrect input.";
+                return terminology;
+            }
+
             await _terminologyRepository.CreateTerminologyAsync(terminology);
 
             return terminology;
@@ -41,6 +63,24 @@ namespace TKDprogress_BLL.Services
 
         public async Task<TerminologyDto> UpdateTerminologyAsync(TerminologyDto terminology)
         {
+            if (terminology.Id <= 0)
+            {
+                terminology.ErrorMessage = "Terminology does not exist.";
+                return terminology;
+            }
+
+            if (terminology.CategoryId <= 0)
+            {
+                terminology.ErrorMessage = "Category type does not exist.";
+                return terminology;
+            }
+
+            if (string.IsNullOrEmpty(terminology.Word) || string.IsNullOrEmpty(terminology.Meaning))
+            {
+                terminology.ErrorMessage = "Terminology name or description has an incorrect input.";
+                return terminology;
+            }
+
             await _terminologyRepository.UpdateTerminologyAsync(terminology);
 
             return terminology;
@@ -48,6 +88,12 @@ namespace TKDprogress_BLL.Services
 
         public async Task<TerminologyDto> DeleteTerminologyAsync(TerminologyDto terminology)
         {
+            if (terminology.Id <= 0)
+            {
+                terminology.ErrorMessage = "Terminology does not exist.";
+                return terminology;
+            }
+
             await _terminologyRepository.DeleteTerminologyAsync(terminology);
 
             return terminology;
